Do not store lint_store in lint contexts
This avoids another case of having a Session and a Guard<LintStore>
simultaneously present in a long-lived structure, which may lead to
deadlocks. In any case, the lint store fields were unused (as best as I
can tell), and are still entirely reachable via the Session if really
needed (though there should be no need).