Remove Session from get_lint_level
We want to avoid having access to Session and LintStore stored in
session at the same time, and callers of this function were one place
where that was happening.

The two things we pulled from there, edition and driver_lint_caps, are
now stored inside the sets directly. That might not quite be ideal, but
it does seem fine from a correctness standpoint.