Stop replacing passes into LintStore
The late, pre_expansion, and early passes always ran at most once during
a given compilation, so simply dropping them once we've run them is
fine.

Ideally, we'd not need to acquire a lock on the lint store at all to
consume these passes -- there's really no reason why we should need to
create the LintPass itself prior to needing it. In an ideal world, we'd
merely have the vector of `Lint`s, and when we go to do pre-expansion
passes, we'd go through the vector and invoke "get pre-expansion pass"
and then run that.

That's a fairly major restructuring though, as we currently register
*passes* not *lints* with the LintStore, which is because multiple lints
may be associated with a single pass. A future commit in this PR will
attempt to clean this up, but this change stands on its own as the
reviewable bit to make sure we can actually drop the passes early.